requirements.txt
# Core Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3

# Selenium and Web Scraping
selenium==4.15.0
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4

# JobSpy - Multi-site job scraper
python-jobspy>=1.1.0

# User Agent Rotation
fake-useragent==1.4.0

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Excel Export
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality (optional, for development)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
